+++ packages/browser-wallet/src/popup/pages/Web3ProofRequest/Display/DisplayBox.tsx
@@ -1,0 +1,40 @@
+import Button from '@popup/shared/Button';
+import Modal from '@popup/shared/Modal';
+import clsx from 'clsx';
+import React, { ReactNode, useState } from 'react';
+import { ClassName } from 'wallet-common-helpers';
+import InfoTooltipIcon from '@assets/svg/info-tooltip.svg';
+
+type DisplayBoxProps = ClassName & {
+    header: string;
+    children: ReactNode;
+    infoBox: ReactNode;
+};
+
+export function DisplayBox({ className, children, header, infoBox }: DisplayBoxProps) {
+    const [open, setOpen] = useState(false);
+
+    return (
+        <section className={clsx('display-box', className)}>
+            <header className="display-box__header">
+                <div className="display6">{header}</div>
+                <Modal
+                    open={open}
+                    onOpen={() => setOpen(true)}
+                    onClose={() => setOpen(false)}
+                    trigger={
+                        <Button clear className="flex">
+                            <InfoTooltipIcon className="display-box__tooltip-icon" />
+                        </Button>
+                    }
+                >
+                    {infoBox}
+                    <Button width="wide" onClick={() => setOpen(false)}>
+                        Ok
+                    </Button>
+                </Modal>
+            </header>
+            {children}
+        </section>
+    );
+}
